Fix "(unknown name)" being the Author for many pyproject.toml packages
No longer does this plugin blindly try to use the Author field only. Now
we're cooking with ~~gas~~ a whole function to merge bare names from
Author with the names (or email addresses, if that's all there is)
provided in Author-Email.

Nearly every one of the dozen-ish "(unknown name)" results I found in my
chat logs now shows an actual author or list of authors, and the only
ones that didn't *actually* don't have author metadata at all.+

This same function is general enough that it can be used to add a
Maintainer field in the output as well, if we'd like to.

----

+ Note that PyPI still attaches a Maintainer section to these packages
  when they're displayed on the web, but that field doesn't appear
  anywhere in the JSON API response so it can't currently be displayed
  by this plugin without serious hax.